test_0: try to get IP of uuid via DNS if any
This is a preparation for the following docker'ization, where each
service runs on a different node(different IP address) via docker.

This tries to get an IP address of UUID via DNS if it's available,
which is the case of docker(-compose) DNS. It falls down to the
original "localhost"(127.0.0.1) if DNS doesn't find an IP of UUID as
it used to be.

Signed-off-by: Hiroshi Doyu <hiroshi.doyu@ericsson.com>